Handling results.pklz file exists but SGE hasn't recognized job completion
Jobs were being unnecessarily rerun because polling for _results.pklz
indicated the job was completed, but when qacct was used to check
the completed status of the job, the job was not recognized as complete.
If this happens, wait 10 seconds and run qacct on the job again to
check for job completion status.